Refactor: 프로필 업로드 리펙토링
